feat: Implement player mechanics including attack, movement, physics, rendering, and stats management
- Added PlayerAttack module for handling player attacks and charge mechanics.
- Introduced PlayerMovement module for vertical movement and collision detection.
- Created PlayerPhysics module for collision and physics helpers.
- Developed PlayerRenderer module for rendering player graphics and animations.
- Implemented PlayerStats module for managing player growth and leveling.
- Added PlayerTimers module for updating various player timers.
- Defined PlayerTypes for memory and input handling.
- Refactored player class to integrate new modules and enhance functionality.
- Updated CollisionSystem and EventSystem to reference the new player structure.
